pull out duplicated scaling logic
